requirements.txt
flask>=3.0.0
Flask-Session>=0.5.0
sqlalchemy>=2.0.29
python-dotenv>=1.0.1
requests>=2.32.0
openai>=1.40.0
gunicorn>=21.2.0
